Simulators/StreamSpeed: Renamed task to StreamVelocity
