+++ src/Simulators/StreamVelocity/StreamGeneratorFactory.hpp
@@ -27,63 +27,57 @@
 // Author: Miguel Aguiar                                                    *
 //***************************************************************************
 
-#ifndef SIMULATORS_STREAM_SPEED_STREAM_GENERATOR_HPP_INCLUDED_
-#define SIMULATORS_STREAM_SPEED_STREAM_GENERATOR_HPP_INCLUDED_
+#ifndef SIMULATORS_STREAM_SPEED_STREAM_GENERATOR_FACTORY_HPP_INCLUDED_
+#define SIMULATORS_STREAM_SPEED_STREAM_GENERATOR_FACTORY_HPP_INCLUDED_
 
-#include <array>
+#include <memory>
+
+#include "DUNE/I18N.hpp"
+
+#include "ModelDataStreamGenerator.hpp"
+#include "StreamGenerator.hpp"
 
 namespace Simulators
 {
-  namespace StreamSpeed
+  namespace StreamVelocity
   {
     namespace StreamGenerator
     {
-      // A simple interface for any sort of source of velocity values.
-      // This base class just provides a constant value.
-      class StreamGenerator
+      //! Factory method for the stream velocity source.
+      //! @param[in] config structure with configuration fields
+      //! @return handle to the stream velocity source.
+      template<typename Configuration>
+      std::unique_ptr<StreamGenerator>
+      factory(Configuration const& config)
       {
-      public:
-        //! Constructor.
-        //! @param[in] wx default stream speed in the North direction (m/s).
-        //! @param[in] wy default stream speed in the East direction (m/s).
-        //! @param[in] wz default stream speed in the Down direction (m/s).
-        StreamGenerator(double wx, double wy, double wz = 0.0);
+        if (config.type == "Constant")
+          return std::make_unique<StreamGenerator>(
+              config.default_wx, config.default_wy, config.default_wz);
+        if (config.type == "Gridded 2D Model Data")
+        {
+          GriddedModelDataConfig mdcfg;
+          mdcfg.filename = config.filename;
+          // Path to the dataset in the file containing the velocity values in
+          // the East direction, given in m/s as u = u(Lat, Lon, Time).
+          mdcfg.u_data_path = "u";
+          // Path to the dataset in the file containing the velocity values in
+          // the North direction, given in m/s as v = v(Lat, Lon, Time).
+          mdcfg.v_data_path = "v";
+          //! Path to the node in the file containing the grid data.
+          //! This node should include the following child nodes:
+          //!   min - array with the lower grid limits.
+          //!   max - array with the upper grid limits.
+          //!   npts - array with the number of points in each dimension.
+          mdcfg.grid_path = "grid";
 
-        ~StreamGenerator() = default;
-
-        //! Get a stream speed value.
-        //! @param[in] lat WGS84 latitude in degrees.
-        //! @param[in] lon WGS84 longitude in degrees.
-        //! @param[in] depth depth in meters.
-        //! @param[in] time time elapsed since the simulation started.
-        //! @return 3-dimensional array containing the stream velocity in the
-        //! North, East and Down directions in meters per second.
-        virtual std::array<double, 3>
-        getSpeed(double lat, double lon, double depth, double time = 0.0) const;
-
-        //! Get the default stream speed.
-        //! @return 3-dimensional array containing the default stream velocity
-        //! value in the North, East and Down directions in meters per second.
-        std::array<double, 3>
-        getDefaultSpeed() const;
-
-        //! Set the default stream speed.
-        //! @param[in] wx default stream speed in the North direction (m/s).
-        //! @param[in] wy default stream speed in the East direction (m/s).
-        //! @param[in] wz default stream speed in the Down direction (m/s).
-        void
-        setSpeed(double wx, double wy, double wz = 0.0);
-
-      private:
-        //! Default speed North.
-        double m_wx;
-        //! Default speed East.
-        double m_wy;
-        //! Default speed Down.
-        double m_wz;
-      };
+          return std::make_unique<Gridded2DModelDataStreamGenerator>(
+              mdcfg, config.default_wx, config.default_wy, config.default_wz);
+        }
+        else
+          throw std::runtime_error(DTR("Unknown stream velocity source type."));
+      }
     }    // namespace StreamGenerator
-  }      // namespace StreamSpeed
+  }      // namespace StreamVelocity
 }    // namespace Simulators
 
 #endif
